implement tryGet for all funs

+++ R/pointDTM100_download.R
@@ -61,7 +61,12 @@ pointDTM100_download = function(voivodeships, outdir = ".", unzip = TRUE, ...) {
 
   for (i in seq_len(nrow(df_names))) {
     filename = paste0(outdir, "/", df_names[i, type], ".zip")
-    utils::download.file(df_names[i, "URL"], filename, mode = "wb", ...)
+    status = tryGet(utils::download.file(df_names[i, "URL"], filename, mode = "wb", ...))
+
+    if (any(status %in% c("error", "warning"))) {
+      return("connection error")
+    }
+
     if (unzip) {
       utils::unzip(filename, exdir = outdir)
       file.remove(filename)
